chore: more migration from `apps.startup` to `jmri.util.startup`
