+++ java/src/apps/PerformScriptModel.java
@@ -1,34 +1,17 @@
 package apps;
 
-import apps.startup.AbstractStartupModel;
-import java.io.File;
-import jmri.JmriException;
-import jmri.script.JmriScriptEngineManager;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 
 /**
  * A PerformScriptModel object runs a script when the program is started.
  *
  * @author Bob Jacobsen Copyright 2003
  * @author Randall Wood (c) 2016
- * @see apps.startup.PerformScriptModelFactory
+ * @see jmri.util.startup.PerformScriptModelFactory
+ * @author Randall Wood Copyright 2020
+ * @deprecated since 4.21.1; use {@link jmri.util.startup.PerformScriptModel} instead
  */
-public class PerformScriptModel extends AbstractStartupModel {
-
-    private final static Logger log = LoggerFactory.getLogger(PerformScriptModel.class);
-
-    public String getFileName() {
-        return this.getName();
-    }
-
-    public void setFileName(String n) {
-        this.setName(n);
-    }
-
-    @Override
-    public void performAction() throws JmriException {
-        log.info("Running script {}", this.getFileName());
-        JmriScriptEngineManager.getDefault().runScript(new File(this.getFileName()));
-    }
+@Deprecated
+@SuppressFBWarnings(value = "NM_SAME_SIMPLE_NAME_AS_SUPERCLASS", justification = "Deprecated by refactoring; retaining unchanged until removal")
+public class PerformScriptModel extends jmri.util.startup.PerformScriptModel {
 }
