+++ java/src/apps/configurexml/PerformFileModelXml.java
@@ -1,9 +1,10 @@
 package apps.configurexml;
 
-import apps.PerformFileModel;
+import jmri.util.startup.PerformFileModel;
 import jmri.util.startup.StartupActionsManager;
 import jmri.InstanceManager;
 import jmri.util.FileUtil;
+
 import org.jdom2.Element;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -13,8 +14,9 @@ import org.slf4j.LoggerFactory;
  *
  * @author Bob Jacobsen Copyright: Copyright (c) 2003
  * @author Ken Cameron Copyright: 2014(c)
- * @see apps.startup.PerformFileModelFactory
+ * @deprecated since 4.21.1; use {@link jmri.util.startup.configurexml.CreateButtonModelXml} instead
  */
+@Deprecated
 public class PerformFileModelXml extends jmri.configurexml.AbstractXmlAdapter {
 
     public PerformFileModelXml() {
