+++ java/src/apps/startup/Bundle.properties
@@ -11,9 +11,9 @@ StartupActionsPreferencesPanel.recommendationsLbl.text=<html>Scripts should be r
 
 # Descriptions of the Startup Models included in JMRI
 apps.CreateButtonModel=Add button to main window
-apps.PerformActionModel=Perform action
-apps.PerformFileModel=Open file
-apps.PerformScriptModel=Run script
+jmri.util.startup.PerformActionModel=Perform action
+jmri.util.startup.PerformFileModel=Open file
+jmri.util.startup.PerformScriptModel=Run script
 apps.startup.TriggerRouteModel=Set route
 apps.startup.ScriptButtonModel=Add script to button
 apps.startup.StartupPauseModel=Pause
