+++ java/src/apps/startup/Bundle_cs.properties
@@ -14,9 +14,9 @@ StartupActionsPreferencesPanel.recommendationsLbl.text=<html>Ve v\u011bt\u0161in
 
 # Descriptions of the Startup Models included in JMRI
 apps.CreateButtonModel=P\u0159idat tla\u010d\u00edtko do hlavn\u00edho okna
-apps.PerformActionModel=Prov\u00e9st akci
-apps.PerformFileModel=Otev\u0159\u00edt soubor
-apps.PerformScriptModel=Spustit skript
+jmri.util.startup.PerformActionModel=Prov\u00e9st akci
+jmri.util.startup.PerformFileModel=Otev\u0159\u00edt soubor
+jmri.util.startup.PerformScriptModel=Spustit skript
 apps.startup.TriggerRouteModel=Nastavit trasu
 apps.startup.ScriptButtonModel=P\u0159idat skript k tla\u010d\u00edtku
 apps.startup.StartupPauseModel=Pozastavit
