+++ java/src/apps/startup/Bundle_da.properties
@@ -11,9 +11,9 @@ StartupActionsPreferencesPanel.recommendationsLbl.text=<html>Scripts skal udf\u0
 
 # Descriptions of the Startup Model included in JMRI
 apps.CreateButtonModel=Tilf\u00f8j en knap til hoved window
-apps.PerformActionModel=Udf\u00f8r en aktion
-apps.PerformFileModel=\u00c5ben en fil
-apps.PerformScriptModel=K\u00f8r et script
+jmri.util.startup.PerformActionModel=Udf\u00f8r en aktion
+jmri.util.startup.PerformFileModel=\u00c5ben en fil
+jmri.util.startup.PerformScriptModel=K\u00f8r et script
 apps.startup.TriggerRouteModel=S\u00e6t en rute
 apps.startup.ScriptButtonModel=Tilf\u00f8j et script til knappen
 apps.startup.StartupPauseModel=Pause
