+++ java/src/apps/startup/Bundle_de.properties
@@ -16,9 +16,9 @@ StartupActionsPreferencesPanel.recommendationsLbl.text=<html>Skripte sollen norm
 
 # Descriptions of the Startup Model included in JMRI
 apps.CreateButtonModel=Platziere Knopf ins Hauptfenster
-apps.PerformActionModel=Handlung durchf\u00fchren
-apps.PerformFileModel=\u00d6ffne Datei
-apps.PerformScriptModel=F\u00fchre Ablaufdatei aus
+jmri.util.startup.PerformActionModel=Handlung durchf\u00fchren
+jmri.util.startup.PerformFileModel=\u00d6ffne Datei
+jmri.util.startup.PerformScriptModel=F\u00fchre Ablaufdatei aus
 apps.startup.TriggerRouteModel=Stelle Weichenstrasse
 EditableStartupAction={0}...
