+++ java/src/apps/startup/ScriptButtonModel.java
@@ -1,17 +1,23 @@
 package apps.startup;
 
+import jmri.util.startup.AbstractStartupModel;
+
 import apps.Apps;
 import apps.gui3.Apps3;
+
 import java.awt.event.ActionEvent;
 import java.io.File;
 import java.io.IOException;
 import java.text.MessageFormat;
+
 import javax.script.ScriptException;
 import javax.swing.AbstractAction;
 import javax.swing.Action;
 import javax.swing.JButton;
+
 import jmri.JmriException;
 import jmri.script.JmriScriptEngineManager;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
