+++ java/src/apps/startup/StartupPauseModel.java
@@ -1,6 +1,8 @@
 package apps.startup;
 
+import jmri.util.startup.AbstractStartupModel;
 import jmri.JmriException;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
