+++ java/src/apps/startup/TriggerRouteModel.java
@@ -1,11 +1,15 @@
 package apps.startup;
 
+import jmri.util.startup.AbstractStartupModel;
+
 import java.util.Locale;
+
 import jmri.InstanceManager;
 import jmri.JmriException;
 import jmri.Route;
 import jmri.RouteManager;
 import jmri.util.prefs.InitializationException;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
