+++ java/src/jmri/util/startup/Bundle_ca.properties
@@ -15,9 +15,9 @@ StartupActionsPreferencesPanel.recommendationsLbl.text=<html>Scripts s'han d'exe
 
 # Descriptions of the Startup Models included in JMRI
 apps.CreateButtonModel=Afegeix un bot\u00f3 a la finestra principal
-apps.PerformActionModel=Realitza l'acci\u00f3
-apps.PerformFileModel=Obre arxiu
-apps.PerformScriptModel=Executa script
+jmri.util.startup.PerformActionModel=Realitza l'acci\u00f3
+jmri.util.startup.PerformFileModel=Obre arxiu
+jmri.util.startup.PerformScriptModel=Executa script
 apps.startup.TriggerRouteModel=Activa ruta
 apps.startup.ScriptButtonModel=Afegeix un script al bot\u00f3
 apps.startup.StartupPauseModel=Atura
