+++ java/src/jmri/util/startup/Bundle_nl.properties
@@ -16,9 +16,9 @@ StartupActionsPreferencesPanel.recommendationsLbl.text=<html>Scripts dienen mees
 
 # Descriptions of the Startup Models included in JMRI
 apps.CreateButtonModel=Voeg knop toe aan hoofdvenster
-apps.PerformActionModel=Voer actie uit
-apps.PerformFileModel=Open bestand
-apps.PerformScriptModel=Voer script uit
+jmri.util.startup.PerformActionModel=Voer actie uit
+jmri.util.startup.PerformFileModel=Open bestand
+jmri.util.startup.PerformScriptModel=Voer script uit
 apps.startup.TriggerRouteModel=Stel Route in
 apps.startup.ScriptButtonModel=Wijs script aan knop toe
 apps.startup.StartupPauseModel=Pause
