+++ java/src/jmri/util/startup/PerformFileModelFactory.java
@@ -1,0 +1,36 @@
+package jmri.util.startup;
+
+
+import jmri.util.startup.StartupModelFactory;
+import jmri.util.startup.StartupModel;
+import jmri.util.startup.PerformFileModel;
+
+import javax.swing.JFileChooser;
+
+import jmri.jmrit.XmlFile;
+
+import org.openide.util.lookup.ServiceProvider;
+
+/**
+ *
+ * @author Randall Wood 2016
+ */
+@ServiceProvider(service = StartupModelFactory.class)
+public class PerformFileModelFactory extends AbstractFileModelFactory {
+
+    @Override
+    public Class<? extends StartupModel> getModelClass() {
+        return PerformFileModel.class;
+    }
+
+    @Override
+    public PerformFileModel newModel() {
+        return new PerformFileModel();
+    }
+
+    @Override
+    protected JFileChooser setFileChooser() {
+        return XmlFile.userFileChooser("XML files", "xml");
+    }
+
+}
