+++ java/src/jmri/util/startup/PerformScriptModelFactory.java
@@ -1,0 +1,39 @@
+package jmri.util.startup;
+
+
+import jmri.util.startup.StartupModelFactory;
+import jmri.util.startup.StartupModel;
+import jmri.util.startup.PerformScriptModel;
+
+import javax.swing.JFileChooser;
+
+import jmri.script.ScriptFileChooser;
+
+import org.openide.util.lookup.ServiceProvider;
+
+/**
+ *
+ * @author Randall Wood
+ */
+@ServiceProvider(service = StartupModelFactory.class)
+public class PerformScriptModelFactory extends AbstractFileModelFactory {
+
+    public PerformScriptModelFactory() {
+    }
+
+    @Override
+    public Class<? extends StartupModel> getModelClass() {
+        return PerformScriptModel.class;
+    }
+
+    @Override
+    public PerformScriptModel newModel() {
+        return new PerformScriptModel();
+    }
+
+    @Override
+    protected JFileChooser setFileChooser() {
+        return new ScriptFileChooser();
+    }
+
+}
