+++ java/src/jmri/util/startup/StartupActionModelUtil.java
@@ -19,7 +19,7 @@ import org.slf4j.LoggerFactory;
 
 /**
  * Maintain a list of actions that can be used by
- * {@link apps.startup.AbstractActionModel} and its descendants. This list is
+ * {@link jmri.util.startup.AbstractActionModel} and its descendants. This list is
  * populated by {@link StartupActionFactory} instances
  * registered with a {@link java.util.ServiceLoader}.
  *
