+++ java/src/jmri/web/server/WebServerPreferencesPanel.java
@@ -4,12 +4,14 @@ package jmri.web.server;
  * @author Steve Todd Copyright (C) 2011
  * @author Randall Wood Copyright (C) 2012, 2014, 2016
  */
-import apps.PerformActionModel;
+import jmri.util.startup.PerformActionModel;
 import jmri.util.startup.StartupActionsManager;
+
 import java.awt.event.ActionEvent;
 import java.awt.event.ItemEvent;
 import java.awt.event.ItemListener;
 import java.util.Arrays;
+
 import javax.swing.GroupLayout;
 import javax.swing.JCheckBox;
 import javax.swing.JComponent;
@@ -18,8 +20,10 @@ import javax.swing.JPanel;
 import javax.swing.JSpinner;
 import javax.swing.LayoutStyle;
 import javax.swing.SpinnerNumberModel;
+
 import jmri.InstanceManager;
 import jmri.swing.PreferencesPanel;
+
 import org.jdesktop.beansbinding.AutoBinding;
 import org.jdesktop.beansbinding.BeanProperty;
 import org.jdesktop.beansbinding.BindingGroup;
