+++ java/test/jmri/util/startup/PerformActionModelFactoryTest.java
@@ -1,0 +1,34 @@
+package jmri.util.startup;
+
+import jmri.util.JUnitUtil;
+
+import org.junit.After;
+import org.junit.Assert;
+import org.junit.Before;
+import org.junit.Test;
+
+/**
+ *
+ * @author Paul Bender Copyright (C) 2017
+ */
+public class PerformActionModelFactoryTest {
+
+    @Test
+    public void testCTor() {
+        PerformActionModelFactory t = new PerformActionModelFactory();
+        Assert.assertNotNull("exists",t);
+    }
+
+    @Before
+    public void setUp() {
+        JUnitUtil.setUp();
+    }
+
+    @After
+    public void tearDown() {
+        JUnitUtil.tearDown();
+    }
+
+    // private final static Logger log = LoggerFactory.getLogger(PerformActionModelFactoryTest.class);
+
+}
